pyproject.toml
[project]
name = "ruleset-generator"
version = "0.1.0"
description = "Generate Surge/Mihomo/Shadowrocket/Loon/sing-box rules from platform-agnostic sources"
readme = "AGENTS.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
]

[tool.uv]
package = false
